[WIP] Add rough draft of public API
The API is documented using RAML[1], a YAML-based language for
describing RESTful APIs.

RAML encourages reuse of resources, so the brunt of the documentation is
contained in the root `resourceTypes` hash.

There are several tools available to render RAML in more accessible
formats. I've bundled an example generated using raml2html[2].

  1. <http://raml.org>
  2. <https://github.com/kevinrenskers/raml2html>